Let's allow methods declared in concrete classes.
Previously, we were only allowing methods declared in abstract classes. While these methods are more likely to be part of the classic skeletal implementation pattern, there should be no reason that we can't use such methods as input to the tool. We'll need to compensate for this decision, however, in #25.